refactor: use single Facility class for all facility types
